chore: run typescript build as part of typecheck, not normal build (#7855)
