+++ package.json
@@ -73,7 +73,9 @@
   },
   "scripts": {
     "build-clean": "rm -rf ./packages/*/build ./packages/*/build-es5",
+    "prebuild": "yarn build:ts",
     "build": "node ./scripts/build.js",
+    "build:ts": "node scripts/buildTs.js",
     "check-copyright-headers": "node ./scripts/checkCopyrightHeaders.js",
     "clean-all": "rm -rf ./node_modules && rm -rf ./packages/*/node_modules && yarn clean-e2e && yarn build-clean",
     "clean-e2e": "find ./e2e -not \\( -path ./e2e/presets/js -prune \\) -not \\( -path ./e2e/presets/json -prune \\) -mindepth 2 -type d \\( -name node_modules -prune \\) -exec rm -r '{}' +",
@@ -84,7 +86,7 @@
     "lint:md": "yarn --silent lint:md:ci --fix",
     "lint:md:ci": "prettylint '**/*.{md,yml,yaml}' --ignore-path .gitignore",
     "postinstall": "opencollective postinstall && yarn build",
-    "publish": "yarn build-clean && yarn build && lerna publish --silent",
+    "publish": "yarn build-clean && yarn typecheck && yarn build && lerna publish --silent",
     "test-ci-es5-build-in-browser": "karma start --single-run",
     "test-ci": "yarn jest-coverage -i --config jest.config.ci.js && yarn test-leak && node scripts/mapCoverage.js && codecov",
     "test-ci-partial": "yarn jest -i --config jest.config.ci.js",
@@ -92,7 +94,8 @@
     "test-leak": "yarn jest -i --detectLeaks jest-mock jest-diff jest-repl",
     "test": "yarn typecheck && yarn lint && yarn jest",
     "typecheck": "flow check --include-warnings",
-    "watch": "yarn build && node ./scripts/watch.js"
+    "watch": "yarn build && node ./scripts/watch.js",
+    "watch:ts": "yarn build:ts --watch"
   },
   "workspaces": {
     "packages": [
